test: rename npsoak suite to soak
Renames test/integration/npsoak -> test/integration/soak and the
package/files to match.
