+++ test/integration/soak/soak_suite_test.go
@@ -1,4 +1,4 @@
-package npsoak
+package soak
 
 import (
 	"context"
